requirements.txt
groq>=0.11.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-docx>=1.1.2
reportlab>=4.2.2
tabulate>=0.9.0
tavily-python>=0.3.0